various project reorg
