+++ pages/api/products/index.ts
@@ -1,0 +1,9 @@
+import {NextApiRequest, NextApiResponse} from "next";
+import {getProducts} from "../../../utils/products";
+
+export default async function handler(req: NextApiRequest, res: NextApiResponse) {
+  if (req.method === 'GET') {
+    const data = await getProducts()
+    res.status(200).json(data);
+  }
+}
